add the edit booking page
